+++ zigbee/clusters/basic.go
@@ -98,7 +98,7 @@ func (b BasicCluster) HandlerAttributes(endpoint zcl.Endpoint, attributes []zcl.
 			for i := 0; i < len(attribute.Value); i++ {
 				switch attribute.Value[i] {
 				case 0x01: // battery voltage
-					bat := float32(zcl.UINT16_(attribute.Value[i+2], attribute.Value[i+3]))
+					bat := float64(zcl.UINT16_(attribute.Value[i+2], attribute.Value[i+3]))
 					i = i + 3
 					b.Ed.Set_battery_params(0, bat/1000)
 
@@ -151,14 +151,14 @@ func (b BasicCluster) HandlerAttributes(endpoint zcl.Endpoint, attributes []zcl.
 					i = i + 5
 
 				case 0x96: // voltage
-					value := float32(uint32(attribute.Value[i+2]) + uint32(attribute.Value[i+3])<<8 + uint32(attribute.Value[i+4])<<16 + uint32(attribute.Value[i+5])<<24)
+					value := float64(uint32(attribute.Value[i+2]) + uint32(attribute.Value[i+3])<<8 + uint32(attribute.Value[i+4])<<16 + uint32(attribute.Value[i+5])<<24)
 					b.Ed.Set_power_source(0x01)
 					b.Ed.Set_mains_voltage(value / 10)
 					fmt.Printf("Voltage:  %0.2fV\n", value/10)
 					i = i + 5
 
 				case 0x97: // current
-					value := float32(uint32(attribute.Value[i+2]) + uint32(attribute.Value[i+3])<<8 + uint32(attribute.Value[i+4])<<16 + uint32(attribute.Value[i+5])<<24)
+					value := float64(uint32(attribute.Value[i+2]) + uint32(attribute.Value[i+3])<<8 + uint32(attribute.Value[i+4])<<16 + uint32(attribute.Value[i+5])<<24)
 					b.Ed.Set_current(value)
 					fmt.Printf("Current: %0.3fA\n", value)
 					i = i + 5
