+++ zigbee/clusters/simple_metering.go
@@ -29,9 +29,9 @@ func (s SimpleMeteringCluster) HandlerAttributes(endpoint zcl.Endpoint, attribut
 				a0000 = true
 				var val64 []byte = append(attribute.Value, 0x0, 0x0)
 				val := binary.LittleEndian.Uint64(val64)
-				s.Ed.Set_energy(float32(val) / 1000)
+				s.Ed.Set_energy(float64(val) / 1000)
 				//CurrentSummationDelivered represents the most recent summed value of Energy, Gas, or Water delivered and consumed in the premises.
-				fmt.Printf("SimpleMeteringCluster attribute id =0x%04x val = %0.2fkWh\n", attribute.Id, float32(val)/100)
+				fmt.Printf("SimpleMeteringCluster attribute id =0x%04x val = %0.2fkWh\n", attribute.Id, float64(val)/100)
 			}
 		default:
 			fmt.Printf("SimpleMeteringCluster unattended attribute id =0x%04x \n", attribute.Id)
